cli: refactor holos render platform command
This patch refactors the holos render platform command to use the new
platform package instead of the previous builder package.  Similar to
the holos show buildplans command, the PerComponentFunc is used, the
main difference is it simply constructs an argument vector to execute
holos render component injecting the proper flags and tags for cue to
produce the build plan.